refactor(canon): split incremental metrics modules
