+++ crates/vize_canon/src/batch/type_checker.rs
@@ -9,7 +9,10 @@ use super::virtual_project::VirtualProject;
 use crate::virtual_ts::{VirtualTsCheckOptions, VirtualTsOptions};
 use vize_carton::String;
 
+mod metrics;
 mod paths;
+mod result;
+pub use metrics::IncrementalCheckMetrics;
 use paths::{IncrementalPaths, collect_project_paths};
 
 /// Result of type checking.
@@ -21,58 +24,6 @@ pub struct TypeCheckResult {
     pub exit_code: i32,
     /// Whether type checking succeeded.
     pub success: bool,
-}
-
-/// Observable work performed by [`TypeChecker::check_incremental`].
-///
-/// Counts are cumulative for one [`BatchTypeChecker`]. `last_*` fields describe
-/// only the most recent incremental call, so embedders and performance gates can
-/// distinguish a reused Corsa session from a silent full-project CLI fallback.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct IncrementalCheckMetrics {
-    /// Number of incremental checks attempted.
-    pub checks: usize,
-    /// Number of Corsa project sessions started successfully.
-    pub session_starts: usize,
-    /// Number of checks served by an existing Corsa project session.
-    pub session_reuses: usize,
-    /// Number of reused sessions that received a non-empty materialized delta.
-    pub session_refreshes: usize,
-    /// Number of incremental checks that degraded from a session to the CLI.
-    pub session_to_cli_fallbacks: usize,
-    /// Whether the most recent incremental check started a Corsa session.
-    pub last_session_started: bool,
-    /// Whether the most recent incremental check reused a Corsa session.
-    pub last_session_reused: bool,
-    /// Whether the most recent incremental check refreshed a reused session.
-    pub last_session_refreshed: bool,
-    /// Whether the most recent incremental check degraded to the CLI.
-    pub last_session_to_cli_fallback: bool,
-    /// Diagnostic input files requested in the most recent incremental check.
-    pub last_requested_files: usize,
-    /// Materialized files whose content changed in the most recent refresh.
-    pub last_changed_files: usize,
-    /// Materialized files created in the most recent refresh.
-    pub last_created_files: usize,
-    /// Materialized files deleted in the most recent refresh.
-    pub last_deleted_files: usize,
-}
-
-impl TypeCheckResult {
-    /// Check if there are any errors.
-    pub fn has_errors(&self) -> bool {
-        self.diagnostics.iter().any(|d| d.severity == 1)
-    }
-
-    /// Get the number of errors.
-    pub fn error_count(&self) -> usize {
-        self.diagnostics.iter().filter(|d| d.severity == 1).count()
-    }
-
-    /// Get the number of warnings.
-    pub fn warning_count(&self) -> usize {
-        self.diagnostics.iter().filter(|d| d.severity == 2).count()
-    }
 }
 
 /// Options for a project-backed batch type checker.
@@ -297,15 +248,6 @@ impl BatchTypeChecker {
         self.project
             .uses_shared_helpers()
             .then_some(crate::virtual_ts::SHARED_PREAMBLE_DTS)
-    }
-
-    /// Return cumulative and most-recent incremental execution metrics.
-    ///
-    /// Reading metrics does not reset them. The cumulative fields remain safe
-    /// under concurrent use; `last_*` fields describe whichever non-empty
-    /// incremental call most recently entered the serialized Corsa session.
-    pub fn incremental_metrics(&self) -> IncrementalCheckMetrics {
-        self.executor.incremental_metrics()
     }
 
     /// Emit declaration files for the scanned project.
